requirements.txt
customtkinter==5.2.1
keyboard==0.13.5
pyperclip==1.8.2
Pillow==10.1.0
psutil==5.9.6
pywin32==306
pystray==0.19.5
